fix!: flat promises
